Introduce episode metadata box in episode fragment
This commit adds infrastructure and UI for episode metadata, in particular media type, playback duration, and download file size,
to be stored, handled, and display by the app. The information in read from the podcast feed and potentially augmented when
the app itself get new (better) data, for example because a download is triggered.